requirements.txt
streamlit
anthropic
pillow